docs: update README and add comprehensive CLI and GUI user guides
- Revamped README.md with detailed architecture layering, feature highlights, and quick-start instructions.
- Added `cli-guide.md` and `gui-guide.md` with screenshots to document command usage and interface features.
- Enhanced the built-in simulator with a default valid CDB structure, application tags, and correct QSFP-DD identifiers.
- Fixed minor type casting in CDB field reading and ensured the simulated bus is opened automatically by the provider.

+++ src/OpenCMIS.CDB.Core/Implementations/CdbReader.cs
@@ -124,8 +124,8 @@ namespace OpenCMIS.CDB.Core
         {
             return type switch
                    {
-                       CdbFieldType.Byte   => value.Length > 0 ? value[0] : 0,
-                       CdbFieldType.Word   => value.Length >= 2 ? (ushort) (value[0] | value[1] << 8) : 0,
+                       CdbFieldType.Byte   => value.Length > 0 ? value[0] : (byte) 0,
+                       CdbFieldType.Word   => value.Length >= 2 ? (ushort) (value[0] | value[1] << 8) : (ushort) 0,
                        CdbFieldType.DWord  => value.Length >= 4 ? (uint) (value[0]   | value[1] << 8 | value[2] << 16 | value[3] << 24) : 0U,
                        CdbFieldType.String => Encoding.ASCII.GetString(value).TrimEnd('\0'),
                        _                   => value
