+++ src/OpenCMIS.UI.WPF/ViewModels/CdbEditorViewModel.cs
@@ -424,8 +424,8 @@ namespace OpenCMIS.UI.WPF.ViewModels
         {
             return type switch
             {
-                CdbFieldType.Byte   => value.Length > 0 ? value[0] : 0,
-                CdbFieldType.Word   => value.Length >= 2 ? (ushort)(value[0] | value[1] << 8) : 0,
+                CdbFieldType.Byte   => value.Length > 0 ? value[0] : (byte)0,
+                CdbFieldType.Word   => value.Length >= 2 ? (ushort)(value[0] | value[1] << 8) : (ushort)0,
                 CdbFieldType.DWord  => value.Length >= 4 ? (uint)(value[0] | value[1] << 8 | value[2] << 16 | value[3] << 24) : 0U,
                 CdbFieldType.String => Encoding.ASCII.GetString(value).TrimEnd('\0'),
                 _                   => (object)value
